ADD
-> difference_days.py file
-> MODIFY deactivate_user.py which will contain the function of same name
-> main.py will now call the two files (difference_days.py and deactivate_user.py) in order to clean code.
